First pass at CDK based Transfer server setup
This is a simplistic transfer server setup to a single bucket. Technically transfer server setups can be quite a bit more complicated, but in our case this isn't needed because we dont have a long lived transfer server need at this time. We only use this service for short lived temporary large file transfers before shutting it down again.